pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "proxyserver"
version = "1.0.0"
description = "Async HTTP/HTTPS Forward Proxy"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "pyyaml>=6.0"
]

[project.scripts]
proxy-server = "run_proxy:main"

[tool.setuptools]
py-modules = ["run_proxy"]
package-dir = {"" = "."}
